Implement report parsing for Tarpaulin
The generated Tarpaulin HTML report does not have stats to parse via the
DOM - the data is injected via a script tag, which only has line-by-line
coverage, though it includes statistics for each file. The total coverage is
counted by summing the covered/coverable stats reported for each file.

Also, the bad version for express is fixed, and messages are
included for errors that occur for the POST endpoint.