Add Metadata core with MongoDB for persistence
With the new process dependency, process handling has been added to
ensure that the ExpressJS server and MongoDB client connections get
closed up properly.

As noted in the Metadata file above the Branch interface, the schema is
definitely not finalized. Eventually metadata will be needed at the repo
level anyways, so reorganizing the document schema is high on the
priority list.