Add template process to serve bash file
Unfortunately, a static file can't really be used to serve the
connection - the host address is needed to link back to the server.
Only a light library is needed to handle the template file though,
and if more are needed in the future it should be a non-trivial
task to add them to the process.

By default I'm configuring this to work with drone.io, since that
is the CI that I personally will be using (most likely). It should
be non-trivial to configure other CI to handle the script as well.