Add Jest to handle unit and integration testing
Wrote a couple of initial unit tests for some of the logic-heavy points
in the application, being logger.ts and formats.ts. Evidently colorize()
causes an error when transforming on the Format object that's returned.
Since that is specific to the Console transport anyways, I just moved it
to the transport, since the unit test was only testing the passed label
in the returned Formats object.

Also the bash template has some issues that I didn't test (oops), so
those are fixed now.